update(api/args): added comments and fixed bugs
Added comments to most of the insane parts of trakt.py/arr.py so its easier to trace the flow. work in progress.

cleaned up redundant/unreachable code in trakt.py,

fixed missing/wrong id's on trakt being repeatedly deleted and readded.

fixed imdb id's that weren't present in arrs not displaying when not found. reverts to trakt's list for it if can't be found in arr.

changed longform arguements to '--'

updated the README with links and stuff

update